% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d3hive.R
\name{renderD3hive}
\alias{renderD3hive}
\title{Widget render function for use in Shiny}
\usage{
renderD3hive(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

